+++ test/unit/test_basic_usages.py
@@ -1,3 +1,7 @@
+from __future__ import annotations
+
+import typing as t
+
 import click
 import pytest
 
@@ -56,6 +60,50 @@ def test_deduce_type_from_nonrequired_nargs_many_argument():
 def test_deduce_type_from_nonrequired_nargs_fixedint_argument():
     arg = click.Argument(["FOO"], nargs=3, required=False)
     assert deduce_type_from_parameter(arg) == (tuple[str, str, str] | None)
+
+
+def test_deduce_type_from_custom_type_with_convert_method():
+    class MyType(click.ParamType):
+        name = "try_int"
+
+        def convert(self, value, param, ctx) -> int | None:
+            try:
+                return int(value)
+            except ValueError:
+                return None
+
+    opt = click.Option(["--foo"], type=MyType())
+    assert deduce_type_from_parameter(opt) == (int | None)
+
+
+def test_deduce_type_from_callback_annotation():
+    def callback(value: t.Any, param: click.Parameter, ctx: click.Context) -> int:
+        try:
+            return int(value)
+        except ValueError:
+            return 0
+
+    opt = click.Option(["--foo"], callback=callback)
+    assert deduce_type_from_parameter(opt) == int
+
+
+def test_type_from_callback_annotation_overrides_custom_type():
+    def callback(value: t.Any, param: click.Parameter, ctx: click.Context) -> int:
+        if isinstance(value, int):
+            return value
+        return 0
+
+    class MyType(click.ParamType):
+        name = "try_int"
+
+        def convert(self, value, param, ctx) -> int | None:
+            try:
+                return int(value)
+            except ValueError:
+                return None
+
+    opt = click.Option(["--foo"], type=MyType(), callback=callback)
+    assert deduce_type_from_parameter(opt) == int
 
 
 def test_check_annotations_fails_on_missing_arg():